requirements.txt
numpy==1.23.2
discord.py
python-dotenv
time
requests